% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_facebook.R
\name{api_facebook_token}
\alias{api_facebook_token}
\title{获取facebook的token}
\usage{
api_facebook_token(key = "353609681364760",
  secret = "1777c63343eba28359537764fab99b9a")
}
\arguments{
\item{key}{key}

\item{secret}{secret}
}
\value{
return value
}
\description{
获取facebook的token
}
\examples{
api_facebook_token();
}
