% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_github.R
\name{api_github}
\alias{api_github}
\title{不需要登录进入github}
\usage{
api_github(path)
}
\arguments{
\item{path}{路径名称}
}
\value{
返回值
}
\description{
不需要登录进入github
}
\examples{
resp <- api_github("/repos/hadley/httr")
}
