% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_github.R
\name{api_github_token}
\alias{api_github_token}
\title{获取github的token设置}
\usage{
api_github_token(key, secret)
}
\arguments{
\item{key}{key}

\item{secret}{secret}
}
\value{
返回值
}
\description{
获取github的token设置
}
\examples{
api_github_token()
}
