% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_google.R
\name{api_google_token}
\alias{api_google_token}
\title{获取google的token}
\usage{
api_google_token(key = "16795585089.apps.googleusercontent.com",
  secret = "hlJNgK73GjUXILBQvyvOyurl")
}
\arguments{
\item{key}{key}

\item{secret}{secret}
}
\value{
返回值
}
\description{
获取google的token
}
\examples{
 api_google_token();
}
