% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_yahoo.R
\name{api_yahoo_token}
\alias{api_yahoo_token}
\title{获取yahoo的token}
\usage{

  api_yahoo_token(key = "dj0yJmk9ZEp0d2J2MFRuakNQJmQ9WVdrOU0zaHRUMlJpTTJNbWNHbzlNQS0tJnM9Y29uc3VtZXJzZWNyZXQmeD00Nw--",
  secret = "82f339a41f71a3b4d9b840be427dde132e36d115")
}
\arguments{
\item{key}{key}

\item{secret}{secret}
}
\value{
return value
}
\description{
获取yahoo的token
}
\examples{
api_yahoo_token();
}
