% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{post_type}
\alias{post_type}
\title{处理post的数据类型}
\usage{
post_type(type = c("json", "form", "query", "header"))
}
\arguments{
\item{type}{数据类型}
}
\value{
返回值
}
\description{
处理post的数据类型
}
\examples{
post_type();
}
