% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_github.R
\name{print.api_github}
\alias{print.api_github}
\title{定义github的定义对象}
\usage{
\method{print}{api_github}(x, ...)
}
\arguments{
\item{x}{api_github}

\item{...}{}
}
\value{
返回值
}
\description{
定义github的定义对象
}
\examples{
print();
}
