% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{timer}
\alias{timer}
\title{Title  timer}
\usage{
timer(count = 100, interval = 1, f)
}
\arguments{
\item{count}{count}

\item{interval}{interval in second}

\item{f}{function}
}
\value{
return
}
\description{
Title  timer
}
\examples{
timer(f=print)
}
