% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_body}
\alias{web2_body}
\title{获取网页的内容到一个向量}
\usage{
web2_body(page, type = "text", encoding = "UTF-8")
}
\arguments{
\item{page}{网页}

\item{type}{类型,默认为text，备用是raw}

\item{encoding}{默认是UTF-8}
}
\value{
返回值
}
\description{
获取网页的内容到一个向量
}
\examples{
web_body()
}
