% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_decode_json}
\alias{web2_decode_json}
\title{将response的json对象进行翻译list}
\usage{
web2_decode_json(page, encoding = "UTF-8")
}
\arguments{
\item{page}{页面对照}

\item{encoding}{编码}
}
\value{
返回值
}
\description{
将response的json对象进行翻译list
}
\examples{
web2_decode_json
}
