% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_login_basic}
\alias{web2_login_basic}
\title{登录窗口}
\usage{
web2_login_basic(url = "http://httpbin.org/basic-auth/user/passwd",
  user = "user", password = "passwd")
}
\arguments{
\item{url}{链接}

\item{user}{用户名}

\item{password}{密码}
}
\value{
返回值
}
\description{
登录窗口
}
\examples{
web2_login_basic
}
