% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_post}
\alias{web2_post}
\title{处理post相应的功能wrapper}
\usage{
web2_post(url = NULL, config = list(), ..., body = NULL,
  encode = c("multipart", "form", "json", "raw"), handle = NULL)
}
\arguments{
\item{url}{网址}

\item{config}{配置}

\item{...}{其他}

\item{body}{表体}

\item{encode}{编码}

\item{handle}{处理}
}
\value{
返回值
}
\description{
处理post相应的功能wrapper
}
\examples{
web2_post();
}
