% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_post_header}
\alias{web2_post_header}
\title{通过头文件传递参数}
\usage{
web2_post_header(url, namedList)
}
\arguments{
\item{url}{post的目标URL}

\item{namedList}{要传入的数据}
}
\value{
返回值
}
\description{
通过头文件传递参数
}
\examples{
web2_post_header();
}
