% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_post_query}
\alias{web2_post_query}
\title{通过参数向url传递数据}
\usage{
web2_post_query(url, namedList)
}
\arguments{
\item{url}{post的目标url}

\item{namedList}{数据}
}
\value{
返回值
}
\description{
通过参数向url传递数据
}
\examples{
web2_post_query();
}
