% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web2.R
\name{web2_testAvgTime}
\alias{web2_testAvgTime}
\title{测试平台连接时间}
\usage{
web2_testAvgTime(url = "http://had.co.nz", path = "index.html",
  times = 20)
}
\arguments{
\item{url}{网址}

\item{path}{首页页地index.html}

\item{times}{连接次数}
}
\value{
返回值
}
\description{
测试平台连接时间
}
\examples{
web2_testAvgTime
}
