% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_getElement}
\alias{web3_getElement}
\title{获取一个或多少element对照}
\usage{
web3_getElement(session, css = NULL, linkText = NULL, partial = NULL,
  xpath = NULL, multiple = FALSE)
}
\arguments{
\item{session}{会话对照}

\item{css}{css选择器}

\item{linkText}{连接文本}

\item{partial}{部分连接文本}

\item{xpath}{xpath}

\item{multiple}{默认单选择FALSE}
}
\value{
返回值
}
\description{
获取一个或多少element对照
}
