% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_session}
\alias{web3_session}
\title{创建一个session会话信息}
\usage{
web3_session(host = "www.reshapedata.com", timeout = 5000,
  mode = "INFO")
}
\arguments{
\item{host}{运行headless web driver host}

\item{timeout}{超时时间}

\item{mode}{调试模式}
}
\value{
返回值
}
\description{
创建一个session会话信息
}
\examples{
web3_session();
}
