% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_session_exit}
\alias{web3_session_exit}
\title{退出session会话}
\usage{
web3_session_exit(session)
}
\arguments{
\item{session}{会议}
}
\value{
返回值
}
\description{
退出session会话
}
\examples{
web3_session_exit();
}
