% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_session_getCurrentUrl}
\alias{web3_session_getCurrentUrl}
\title{返回当前的URL文本信息}
\usage{
web3_session_getCurrentUrl(session)
}
\arguments{
\item{session}{会话信息}
}
\value{
返回值
}
\description{
返回当前的URL文本信息
}
\examples{
web3_session_getCurrentUrl();
}
