% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_session_goto}
\alias{web3_session_goto}
\title{session操作goto转向某个URL}
\usage{
web3_session_goto(session, url = NULL)
}
\arguments{
\item{session}{会话}

\item{url}{地址}
}
\value{
返回值
}
\description{
session操作goto转向某个URL
}
\examples{
web3_session_goto();
}
