% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_data}
\alias{web_data}
\title{获取节点数据}
\usage{
web_data(web_page, node_selector = ".el", attr_name = "href",
  node_loc = TRUE)
}
\arguments{
\item{web_page}{网页}

\item{node_selector}{节点选择器}

\item{attr_name}{属性名}

\item{node_loc}{节点还是属性}
}
\value{
返回值
}
\description{
获取节点数据
}
\examples{
 web_data()
}
