% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_page}
\alias{web_page}
\title{获取web网页信息}
\usage{

  web_page(url = "https://search.51job.com/list/020000,000000,0000,00,9,99,\%25E6\%2595\%25B0\%25E6\%258D\%25AE\%25E7\%25A7\%2591\%25E5\%25AD\%25A6,2,1.html?lang=c&stype=1&postchannel=0000&workyear=99&cotype=99&degreefrom=99&jobterm=99&companysize=99&lonlat=0\%2C0&radius=-1&ord_field=0&confirmdate=9&fromType=4&dibiaoid=0&address=&line=&specialarea=00&from=&welfare=")
}
\arguments{
\item{url}{输入网页地址}
}
\value{
返回值
}
\description{
获取web网页信息
}
\examples{
web_page()
}
